Исправление замечаний
