+++ DocumentModule.Infrastructure/FileStorage/FileStorageSettings.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DocumentModule.Infrastructure.FileStorage
+{
+    public class FileStorageSettings
+    {
+        public string Endpoint { get; set; }
+        public string AccessKey { get; set; }
+        public string SecretKey { get; set; }
+        public string BucketName { get; set; }
+    }
+}
